requirements.txt
Pyrogram==2.0.106
python-dotenv==1.1.1
pybit==5.11.0